+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/MetadataQueryService.java
@@ -104,7 +104,7 @@ public class MetadataQueryService implements org.apache.skywalking.oap.server.li
     }
 
     @SneakyThrows
-    public Service getService(final String serviceId) throws IOException {
+    public Service getService(final String serviceId) {
         final List<Service> services = this.combineServices(this.serviceCache.get(true).get(serviceId));
         return CollectionUtils.isNotEmpty(services) ? services.get(0) : null;
     }